Updating editing a claim

+++ app/Http/Controllers/AdjusterController.php
@@ -551,38 +551,40 @@ class AdjusterController extends Controller
                     }
                 }
                 $documents = Document::where(["claimID" => $claim->id])->get();
-                if (count($documents) > 0) {
-                    $affectedDocumentRows = Document::where(["claimID" => $claim->id])->delete();
-                    if ($affectedDocumentRows > 0) {
-                        foreach ($documents as $document) {
-                            $image_path = "documents/" . $document->name;  // Value is not URL but directory file path
-                            if (File::exists($image_path)) {
-                                File::delete($image_path);
+                if(isset($totalImages)) {
+                    if (count($documents) > 0) {
+                        $affectedDocumentRows = Document::where(["claimID" => $claim->id])->delete();
+                        if ($affectedDocumentRows > 0) {
+                            foreach ($documents as $document) {
+                                $image_path = "documents/" . $document->name;  // Value is not URL but directory file path
+                                if (File::exists($image_path)) {
+                                    File::delete($image_path);
+                                }
                             }
                         }
                     }
-                }
-                for ($x = 0; $x < $totalImages; $x++) {
-                    if ($request->hasFile('images' . $x)) {
-                        $file = $request->file('images' . $x);
-                        $filename = $file->getClientOriginalName();
-                        $extension = $file->getClientOriginalExtension();
-                        $path = $file->getRealPath();
-                        $size = $file->getSize();
-                        $picture = date('His') . '-' . $filename;
-                        //Save files in below folder path, that will make in public folder
-                        $file->move(public_path('documents/'), $picture);
-                        $documents = Document::create([
-                            "claimID" => $claimID,
-                            "name" => $picture,
-                            "mime" => $extension,
-                            "size" => $size,
-                            "documentType" => Config::$DOCUMENT_TYPES["IMAGE"]["ID"],
-                            "url" => $path,
-                            "segment" => Config::$ASSESSMENT_SEGMENTS["ASSESSMENT"]["ID"],
-                            "createdBy" => Auth::id(),
-                            "dateCreated" => $this->functions->curlDate()
-                        ]);
+                    for ($x = 0; $x < $totalImages; $x++) {
+                        if ($request->hasFile('images' . $x)) {
+                            $file = $request->file('images' . $x);
+                            $filename = $file->getClientOriginalName();
+                            $extension = $file->getClientOriginalExtension();
+                            $path = $file->getRealPath();
+                            $size = $file->getSize();
+                            $picture = date('His') . '-' . $filename;
+                            //Save files in below folder path, that will make in public folder
+                            $file->move(public_path('documents/'), $picture);
+                            $documents = Document::create([
+                                "claimID" => $claimID,
+                                "name" => $picture,
+                                "mime" => $extension,
+                                "size" => $size,
+                                "documentType" => Config::$DOCUMENT_TYPES["IMAGE"]["ID"],
+                                "url" => $path,
+                                "segment" => Config::$ASSESSMENT_SEGMENTS["ASSESSMENT"]["ID"],
+                                "createdBy" => Auth::id(),
+                                "dateCreated" => $this->functions->curlDate()
+                            ]);
+                        }
                     }
                 }
                 $response = array(
